+++ src/ProblemDetailsMiddlewareFactory.php
@@ -1,7 +1,7 @@
 <?php
 /**
  * @see       https://github.com/zendframework/zend-problem-details for the canonical source repository
- * @copyright Copyright (c) 2017 Zend Technologies USA Inc. (https://www.zend.com)
+ * @copyright Copyright (c) 2017-2018 Zend Technologies USA Inc. (https://www.zend.com)
  * @license   https://github.com/zendframework/zend-problem-details/blob/master/LICENSE.md New BSD License
  */
 
@@ -15,8 +15,6 @@ class ProblemDetailsMiddlewareFactory
 {
     public function __invoke(ContainerInterface $container) : ProblemDetailsMiddleware
     {
-        return $container->has(ProblemDetailsResponseFactory::class)
-            ? new ProblemDetailsMiddleware($container->get(ProblemDetailsResponseFactory::class))
-            : new ProblemDetailsMiddleware();
+        return new ProblemDetailsMiddleware($container->get(ProblemDetailsResponseFactory::class));
     }
 }