+++ test/ProblemDetailsResponseFactoryFactoryTest.php
@@ -10,10 +10,14 @@ declare(strict_types=1);
 namespace ZendTest\ProblemDetails;
 
 use Closure;
+use PHPUnit\Framework\Assert;
 use PHPUnit\Framework\TestCase;
 use Psr\Container\ContainerInterface;
 use Psr\Http\Message\ResponseInterface;
-use Zend\Diactoros\Response;
+use ReflectionProperty;
+use RuntimeException;
+use stdClass;
+use TypeError;
 use Zend\ProblemDetails\ProblemDetailsResponseFactory;
 use Zend\ProblemDetails\ProblemDetailsResponseFactoryFactory;
 
@@ -24,11 +28,48 @@ class ProblemDetailsResponseFactoryFactoryTest extends TestCase
         $this->container = $this->prophesize(ContainerInterface::class);
     }
 
-    public function testLackOfOptionalServicesResultsInFactoryUsingDefaults() : void
+    public function assertResponseFactoryReturns(ResponseInterface $expected, ProblemDetailsResponseFactory $factory)
+    {
+        $r = new ReflectionProperty($factory, 'responseFactory');
+        $r->setAccessible(true);
+        $responseFactory = $r->getValue($factory);
+
+        Assert::assertSame($expected, $responseFactory());
+    }
+
+    public function testLackOfResponseServiceResultsInException()
+    {
+        $factory = new ProblemDetailsResponseFactoryFactory();
+        $e = new RuntimeException();
+
+        $this->container->has('config')->willReturn(false);
+        $this->container->get('config')->shouldNotBeCalled();
+        $this->container->get(ResponseInterface::class)->willThrow($e);
+
+        $this->expectException(RuntimeException::class);
+        $factory($this->container->reveal());
+    }
+
+    public function testNonCallableResponseServiceResultsInException()
+    {
+        $factory = new ProblemDetailsResponseFactoryFactory();
+
+        $this->container->has('config')->willReturn(false);
+        $this->container->get('config')->shouldNotBeCalled();
+        $this->container->get(ResponseInterface::class)->willReturn(new stdClass);
+
+        $this->expectException(TypeError::class);
+        $factory($this->container->reveal());
+    }
+
+    public function testLackOfConfigServiceResultsInFactoryUsingDefaults() : void
     {
         $this->container->has('config')->willReturn(false);
-        $this->container->has(ResponseInterface::class)->willReturn(false);
-        $this->container->has('Zend\ProblemDetails\StreamFactory')->willReturn(false);
+
+        $response = $this->prophesize(ResponseInterface::class)->reveal();
+        $this->container->get(ResponseInterface::class)->willReturn(function () use ($response) {
+            return $response;
+        });
 
         $factoryFactory = new ProblemDetailsResponseFactoryFactory();
         $factory = $factoryFactory($this->container->reveal());
@@ -41,8 +82,8 @@ class ProblemDetailsResponseFactoryFactoryTest extends TestCase
             $factory
         );
 
-        $this->assertAttributeInstanceOf(Response::class, 'response', $factory);
-        $this->assertAttributeInstanceOf(Closure::class, 'bodyFactory', $factory);
+        $this->assertAttributeInstanceOf(Closure::class, 'responseFactory', $factory);
+        $this->assertResponseFactoryReturns($response, $factory);
     }
 
     public function testUsesDebugSettingFromConfigWhenPresent() : void
@@ -50,8 +91,8 @@ class ProblemDetailsResponseFactoryFactoryTest extends TestCase
         $this->container->has('config')->willReturn(true);
         $this->container->get('config')->willReturn(['debug' => true]);
 
-        $this->container->has(ResponseInterface::class)->willReturn(false);
-        $this->container->has('Zend\ProblemDetails\StreamFactory')->willReturn(false);
+        $this->container->get(ResponseInterface::class)->willReturn(function () {
+        });
 
         $factoryFactory = new ProblemDetailsResponseFactoryFactory();
         $factory = $factoryFactory($this->container->reveal());
@@ -66,47 +107,13 @@ class ProblemDetailsResponseFactoryFactoryTest extends TestCase
         $this->container->has('config')->willReturn(true);
         $this->container->get('config')->willReturn(['problem-details' => ['json_flags' => JSON_PRETTY_PRINT]]);
 
-        $this->container->has(ResponseInterface::class)->willReturn(false);
-        $this->container->has('Zend\ProblemDetails\StreamFactory')->willReturn(false);
+        $this->container->get(ResponseInterface::class)->willReturn(function () {
+        });
 
         $factoryFactory = new ProblemDetailsResponseFactoryFactory();
         $factory = $factoryFactory($this->container->reveal());
 
         $this->assertInstanceOf(ProblemDetailsResponseFactory::class, $factory);
         $this->assertAttributeSame(JSON_PRETTY_PRINT, 'jsonFlags', $factory);
-    }
-
-    public function testUsesResponseServiceFromContainerWhenPresent() : void
-    {
-        $response = $this->prophesize(ResponseInterface::class)->reveal();
-
-        $this->container->has('config')->willReturn(false);
-        $this->container->has(ResponseInterface::class)->willReturn(true);
-        $this->container->get(ResponseInterface::class)->willReturn($response);
-        $this->container->has('Zend\ProblemDetails\StreamFactory')->willReturn(false);
-
-        $factoryFactory = new ProblemDetailsResponseFactoryFactory();
-        $factory = $factoryFactory($this->container->reveal());
-
-        $this->assertInstanceOf(ProblemDetailsResponseFactory::class, $factory);
-        $this->assertAttributeSame($response, 'response', $factory);
-    }
-
-    public function testUsesStreamFactoryServiceFromContainerWhenPresent() : void
-    {
-        // @codingStandardsIgnoreStart
-        $streamFactory = function () { };
-        // @codingStandardsIgnoreEnd
-
-        $this->container->has('config')->willReturn(false);
-        $this->container->has(ResponseInterface::class)->willReturn(false);
-        $this->container->has('Zend\ProblemDetails\StreamFactory')->willReturn(true);
-        $this->container->get('Zend\ProblemDetails\StreamFactory')->willReturn($streamFactory);
-
-        $factoryFactory = new ProblemDetailsResponseFactoryFactory();
-        $factory = $factoryFactory($this->container->reveal());
-
-        $this->assertInstanceOf(ProblemDetailsResponseFactory::class, $factory);
-        $this->assertAttributeSame($streamFactory, 'bodyFactory', $factory);
     }
 }